pyproject.toml
[project]
name = "ai-agent-platform"
version = "0.2.0"
description = "Multi-agent AI platform with LangGraph orchestration, MCP tools, and multi-modal I/O"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.15.0",
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.13.0",
    "ddgs>=9.0.0",
    "fastapi>=0.135.1",
    "fpdf2>=2.8.0",
    "httpx>=0.28.0",
    "langchain>=1.2.12",
    "langchain-openai>=0.3.0",
    "langgraph>=1.1.1",
    "mcp>=1.0.0",
    "openai>=1.60.0",
    "openpyxl>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "redis>=7.3.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "sse-starlette>=2.0.0",
    "uvicorn>=0.41.0",
]

[project.optional-dependencies]
anthropic = ["langchain-anthropic>=0.3.0"]
google = ["langchain-google-genai>=2.1.0"]
ollama = ["langchain-ollama>=0.3.0"]
groq = ["langchain-groq>=0.3.0"]
all-llm = [
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.1.0",
    "langchain-ollama>=0.3.0",
    "langchain-groq>=0.3.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "httpx>=0.28.0",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.15.5",
]
